+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+### Changed
+- moved biggest part of bootstrap cleanup into new img_build.py script
 
 ## [0.3.0] - 2018-06-27
 ### Added